requirements.txt
fastapi
uvicorn
requests
PyMuPDF
google-generativeai
pinecone
openai
python-dotenv